refactor to expose supabase types
